Rewrite binary recursive descent parser into state machine
Previously the binary tape parser was a recursive descent parser, but it
became too easy to overflow the stack given a heavily nested object. To
avoid stack overflows, the parser was rewritten to use a state machine.
This also aligns the parser to simdjson's stage 2. The hope was to net
some performance gains here, but none were achieve according to the
jomini and web benchmarks