Move tape stack tracking into token tape
Previously there was a separate state tracking array alongside a parser's token
tape that kept track of the transition step after an object or array was
terminated. Every level deeper into the data (a nested object, array, etc)
would cause another level of the array to be consumed. This array had a limit.
Initially this limit was justified for performance reasons and then as a
countermeasure for nefarious input as I did not see a use case for heavily
nested input. First the limit 15 was levels deep, but then revised to 30 to
accept once it recursive events were discovered.

A recursive event is event whose next state is dependent on previous actions.
The [Oxford Symposium Event
(flavor_eng.9880)](https://eu4.paradoxwikis.com/English_events#flavor_eng.9880)
is a great example of this. Every time this event fires, a new nested object is
created. Here's an example from one of my playthroughs as great britain:

```
delayed_event={
 event="flavor_eng.9880"
 days=3199
 scope={
  country="GBR"
  scope_is_valid=yes
  seed=57955.08.26
  random=57955.08.26
  from={
   country="GBR"
   scope_is_valid=yes
   seed=13753.12.05
   random=13753.12.05
   from={
    country="GBR"
    scope_is_valid=yes
    seed=18494.05.24
    random=18494.05.24
    from={
     country="GBR"
     scope_is_valid=yes
     seed=3491.04.17
     random=3491.04.17
    }
   }
  }
 }
}
```

The only practical limit to how many levels deep this event can be nested in
game time. If the event fires every 10 years, then it should be impossible for
this event to fire more than (1821 - 1444) / 10 = 37 times. This is when I
raised the limit from 15 to 30 as someone had exceeded 150 years of symposiums.
30 was chosen as I did not believe one could reasonably achieve more than 300
years of symposiums.

This all changed when I saw that there are EU4 mods that create recursive
events that fire every year instead of every decade. Now all of sudden this
could create data that is nested more than 300 times. Something had to change.
While Rakaly doesn't cater to mods, Rakaly should be able to at least parse
these saves.

The obvious fix would be to swap out a fixed length array for a growable one.
This worked but it came with a small but tangible performance and memory cost.

Then an idea came to me. The token tape used for parsing contains a start and
end pointer for each array and object (this way one can quickly hop around an
object). However until the end of the object / array is reached, instead of
storing a blank for the end pointer, store a pointer to the start of the parent
object / array. This way when an object / array ends, we look up the its
parent's type before filling in the end pointer with the correct value.

Now there is no limit to how deep events can be nested and previous state
tracking logic could be removed. This is an instance of where more code was
removed than added to implement this feature.