Consolidate BinaryTape data into a single tape
Previously the BinaryTape would split out text and rbg data into their
separate tapes and BinaryToken::Text would reference that index. The
thought was this would minimize the size of BinaryTape to 16 bytes
instead of 24 bytes if a scalar was injected (and since Rgb values are
so rare, it doesn't matter where that is put).

The issue with this design is that it is not ergonomic. Yes, jomini is
not supposed to be a super ergonomic layer but feels better as a user to
have all the data necessary in the Token instead of being directed
elsewhere.

Benchmarks showed no change when scalars were embedded directly in a
BinaryToken. This also makes BinaryToken and TextToken symmetrical in
their lifetime parameters.

Given all this, BinaryToken should have the data tape and rgb tape
consolidated with the token tape