Integrate tape clear into tape parse
Previously it was error prone to remember to call clear before a call to
parse. I couldn't find a use case where one wouldn't want to call clear
for every parse, so now a tape is cleared on every parse.

Also according to API guidelines, if an object provides a `Default`
implementation, it should also include a `new` constructor.