Scalar to_utf8 trims trailing whitespace
EU4 binary files contains a `flags` section. In this section, the values
contain a trailing newline. This is a marked difference from the text
format where these values don't end with a newline. In order for both
binary and text formats to deserialize to the same value, there needs to
be some intermediate layer. In EU4save, this was done at the serde layer
for EU4 dates (as this is the only case in EU4 saves where the binary
value differs from text value). By trimming whitespace for all scalars
converted to strings, we avoid this for all situations.

I can't imagine the value add for keeping around the newline on purpose
(but if desired, can be seen in `view_data`)

This commit introduces the precedent of modifying the input to a scalar.

Benchmarks show that trimming the end from strings is a non-negligle
operation for short strings that don't need to be heap allocated
(`to_utf8()` (not `to_utf8_owned()`)). Time to decode strings of length
1 to utf-8 increased 36%. This penalty decreases the longer the string.
Latency for strings of length 4 increased 10%.