Introduce the concept of a binary flavor (#18)
CK3 and EU4 seem to have a share a binary format that differs only in how rational (ie: not whole) numbers should be decoded.

While jomini can be used to parse HOI4 and Imperator saves, I have not yet run the numbers to ensure the values are being decoded correctly, so we'll assume, for the time being, that HOI4, Imperator, and EU4 all share the same format and that it is CK3 that introduces the change.

There are two tokens that signal a rational number is to be decoded next:

The first one is `0x000d`:

 - EU4 stores the value as a whole integer by multiplying the value by 1000. [So to follow berkley's lead](https://inst.eecs.berkeley.edu/~cs61c/sp06/handout/fixedpt.html), this is a `fixed<32,3>` value (32 bits wide, with 3 **decimal digit** (not binary) positions in the number)
 - CK3 decodes directly as a IEEE-754 floating point value (ie: `f32`)

The second one is `0x0167`:

- EU4 decodes it as a [Q17.15](https://en.wikipedia.org/wiki/Q_(number_format)) with 5 fractional digits
- CK3 decodes this how EU4 decodes `0x000d`

So this PR introduces the `BinaryFlavor` trait to `BinaryTape` (and consequentially to `BinaryDeserializer`) to know how to derive rational values from the data. This makes it easier to share the parsing code while allowing some extensibility while still maintaining [top tier performance (as the tape parser is monomorphized to the trait's implementation)](https://doc.rust-lang.org/book/ch10-01-syntax.html#performance-of-code-using-generics).

This PR also renames the `F32` and `Q16` binary tokens as they could be considered misnomers now. Since there isn't really distinctive nature of the two rational indicating tokens that the parser decodes into 32bit floating point values, I've named them `F32_1` and `F32_2`.