Reject nested values inside hidden objects
While hidden objects are supported:

```
levels={10 0=1 2=3}
```

What is not supported is if the hidden objects contains anything other than
atomic values:

```
levels={10 0=1 2={3 4}}
```

I've yet to see container values inside a hidden object so this should
not be limiting. I'm introducing this limitation to ease logic in
downstream projects ck3save and eu4save.