Major binary parsing performance improvements
Benchmarks show 30-80% increase in throughput (we can consistently
exceed 1.2 GiB/s now). Key insights:

- Having `BinaryToken::Rgb` be boxed (ie heap allocated) meant that a
`BinaryToken` needs to be dropped causing all sorts of slowdowns
(swapping values and clearing the vector would take 20% of the parse
time as the value would need to be dropped)
- Furthermore, inlining `Rgb` in the TextToken did not increase the size
of the enum (so it should have never been boxed in the first place).
- Rewrote parsing state machine to determine the token first before
looking at the state. When parsing scalar data, the next state can be
trivially determined