Add common game date implementation
Three downstream projects use a common date implementation:

 - eu4
 - ck3
 - imperator

Copying and pasting this code to every project is error prone,
especially when there are bugfixes, features, or performance changes
made in one, it has to be replicated to the rest.

So I decided to create a common module to store data structures that may
be common across games.

It should be noted that HOI4 saves use hours and the binary format
includes hours (this implementation just ignores it), so there is a case
for adding hour data. This commit doesn't as I've yet to decide how to
best to ensure that games that can't represent hours use data structures
that don't have hours. More pondering is necessary.