Add operator text token
This adds supports for parsing non-equal operators from text data:

```
val < 3
val <= 3
val > 3
val >= 3
```

This is done by pushing a `TextToken::Operator` into the tape. Note that
this operator will only appear if an operator is present and is
non-equal. Since equals account for all operators in saves, and saves
have millions of tokens, potentially adding 1/3 more tokens is dangerous
for performance and memory usage for very little gain.

Benchmarks show this change created a low single digit performance
degradation but I find this tradeoff worth it.  I believe this to be the
last step in supporting the full syntax of text files so that at least
some value can be derived from every file.