Expose hidden objects as Text / Binary token
To recap, hidden objects (aka as object headers) look like:

```
a = { 10 0=1 1=2 }
```

It's currently parsed as equivalent to an array of two values (`10` and
an object):

```
a = { 10 { 0=1 1=2 } }
```

The current API doesn't expose to the client if the object they are
looking at has been hidden inside an array. This is important to melters
(binary -> plaintext) as we want to be able to preserve the correct
structure in the conversion.

To fix this, `BinaryToken::HiddenObject` and `TextToken::HiddenObject`
were introduced. They behave exactly like regular objects except that
they denote the object is hidden.