Add support for generic token headers
Given the format of

```
<key> = <header> { <data> }
```

This PR introduces the support for arbitrary header values. Previously
only `rgb` was supported for text parsing. This was made generic to
allow for other formats such as:

```
color = hsv { 0.58 1.00 0.72 }
color = rgb { 169 242 27 }
color = hsv360{ 25 75 63 }
position = cylindrical{ 150 3 0 }
color5 = hex { aabbccdd }
mild_winter = LIST { 3700 3701
    # ...
}
```
Since new headers can be introduced whenever, it may not make sense to
hard code in the entire list as then then there would be too many text
tokens related to this syntax. I want to avoid `TextToken::{Rgb, Hsv,
Hsv360, ...}` as that seems like pollution at the lowest level. So
`TextToken::Rgb` was removed in favor of `TextToken::Header("rgb")`. As
can imagine a header token describes the subsequent token in the tape.

The binary parser keeps its Rgb token as that is the only token that has
thus far appeared in binary data.

Another factor that has changed is deserialization. Previously Rgb for
both text and binary would deserialize as a 3 element integer sequence
but with the concept of headers (and the goal that one should be able to
use the same deserializer for both binary and text), this has changed so
that they yield the header as the first element followed by the data
element.