Skip empty containers at start of object
This commit allows the following data to be parsed both in text and binary form.

```
history = {
  {}
  1689.10.2={
    decision="abc123"
  }
}
```

This format can be seen in some EU4 saves. What causes this is unknown.

I believe this to be a regression in behavior. At one point this format could
be parsed, but must have been removed by CK3's introduction of the following
format:

```
levels = { 10 0=1 1=2}
```

This PR allows both formats to be parsed and adds test cases so we don't regress.

Benchmarks showed minor degredations for the binary format and minor
improvements for text. Not enough for me to investigate further.