Give hint to deserializer about length of values / fields
When deserializing into a vector or a hashmap, the deserializer will see
if there is a hint available so that it can pre-allocate a vector of the
right size. This means that there is no penalty for needing to grow the
vector. Overall this shaved off 4-5% in eu4save's debug_save benchmark.