Support negative dates
It turns out that all EU4 saves contain a monarch with a negative date:

```
monarch={
    id={
      id=12
      type=48
    }
    name="Pietro I"
    country="ACH"
    DIP=1
    ADM=1
    MIL=2
    culture=lombard
    religion=catholic
    dynasty="de Saint-Superan"
    birth_date=-17.1.1
}
```

The negative dates are also encoded in the binary format. This breaks
the date detection. With the new found knowledge that dates can be
negative, the code has been updated so that years are stored as signed
integers.

This is /technically/ a breaking change but I believe this is more in
the bugfix category.