Allow scalar to float to fail with loss of precision
In CK3 and imperator saves (at the very least) there are 64 bit unsigned
integers that can't be represented perfectly as 64 bit floating point:

```
identity=18446744073709351616
```

However languages will allow one to cast to floating point and sacrifice
precision and interpret the value as:

```
18446744073709548000
```

This poses a problem. While the two values may be numerically equivalent
in the eyes of floating point, they should not be considered equivalent
when distinguishing objects.

For use cases where one knows ahead of time that a value is 64 bit
unsigned, there is no issue. It's the use cases where one is traversing
the document and trying to determine types (eg: converting to JSON). To
save users from the footgun of silently losing precision, the
`scalar.to_f64` will now fail when an integer can't be represented
accurately.

The good news is that the old behavior of accepting precision loss can
be preserved, as the `ScalarError::PrecisionLoss` contains the float
value with the precision loss.

The major tradeoff here is that the identifiers that fall within the
safe range could be parsed as 64 bit floating point while keys of the
same name that don't fall within the safe range would be a string. This
might seem odd (two keys of the same name, one being able to be parsed
as a number while the other one is not) it is more easy to work with.