An incomplete object should fail to parse
Previously `T&}` was allowed to be parsed into a tape of a single key.
This subsequently broke the reader as every object reader assumes that
there is a value for every key and so would panic on this input.

I believe this possibility of panicing was introduced in 0.9.1 with
the addition of vic2 support which required allowing a trailing brace