Support interpolated variables
Interpolated variables look like:

```
@[1-leopard_x]
```

Where the above is an expression to be evaluated by the game. This
syntax can be found in the ck3 game files.

Previously the parser would detect the above expression as multiple
scalars. Now the parser detects it as one unquoted scalar. Downstream
users will still need to detect if a scalar is a variable and how to
evaluate the expression. Some day I may add in an additional
`TextToken::Variable` in some capacity.

This commit passed fuzzing and benchmarks showed negligible change.