BinaryFlavor receives arrays instead of slices for arguments
This should make it more intuitive what the inputs of the function are
without needing to read documentation. The implementation uses a
simplified and const generic version of arrayref, which helped reduce
the number of unsafe usages significantly.

Benchmarks showed no regression when this was implemented so
preferring arrays to unaligned slice pointer reads was applied across
the code base