Rename secondary binary float point encoding to f64
There's a bit of an oddity with the binary tokens. There are two
floating point values:

```rust
BinaryToken::F32_1(f32)
BinaryToken::F32_2(f32)
```

The second version actually consumes 8 bytes of data, so a better design
should be:

```rust
BinaryToken::F64(f64)
```

Before #67, only 4 bytes of data was looked at for this second version.
This led to incorrect decoding as seen by the eu4save's burg.eu4 and
cologne2.eu4 which when only 4 bytes were considered the
`total_military_power` field was decoded as negative instead of a large
positive.