Remove cap on negative years in dates
With the release of leviathan we now have prehistoric monuments with
large negative dates (defined arbitrarily as a magnitude greater than
-100):

```
-2500.1.1
```

These dates will fail to parse in both text and binary forms as years
below -100 were considered invalid due to binary integers and dates
being encoded the same way.

The fix is to allow constructing a date with an unbounded negative date,
but to introduce a function that allows binary format users to ask if
the data is more likely a date than an integer.