Skip all empty objects
In 1.31.2 EU4 saves there is syntax like:

```plain
history={
  capital=945 {}
  1444.11.11={
    # ..
  }
}
```

Previously the `945` would be parsed as the header to an empty array
(reminder that headers are those like `rgb`, `hsv`, `hsv360`, `LIST`).

This commit recognizes it as an empty object and skips it.

Consequently the parser state machine could be simplified leading to a
10-20% throughput increase in benchmarks