Rewrite Scalar methods to take self instead of reference
Clippy flagged all the scalar methods as breaking the wrong self
convention: https://rust-lang.github.io/rust-clippy/master/index.html#wrong_self_convention

Following the logic from: https://rust-lang.github.io/api-guidelines/naming.html#ad-hoc-conversions-follow-as_-to_-into_-conventions-c-conv

> `f64::to_radians()` converts a floating point quantity from degrees to
radians. The input is `f64`. Passing a reference `&f64` is not warranted
because `f64` is cheap to copy. Calling the function `into_radians` would be
misleading because the input is not consumed.

Since scalar is `Copy` it doesn't warrant a `self` reference.

Initially I was considering renaming them to `as_*` but realized that is
inappropriate as the conversion is not free (can cost up to 10 ns to
perform the slowest conversion), and `as_*` methods are considered to be
free.