Support decoding numbers with leading plus sign
It appears that the game engine allows for explicit positive numbers:

In EU4 golden bulls:

```
yearly_absolutism = +0.5
```

In Stellaris static modifiers:

```
pop_happiness = +0.10
```

Did not find this occurence in HOI4, Imperator, or CK3.

This commits allows a plus sign to be present (and skipped) when
decoding a number from a scalar. Benchmarks show no performance changes.