Fix binary parse for deep object
The binary parser would get confused whenever it saw in quick
succession: two array starts and then an object start as it would think
the object was hidden. This caused the HOI4 downstream melter (where
this situation occurs with `operatives`) to fail with unmatched curly
braces.

The solution is admittedly a bit of a kludge and corrects the state of
the binary parser by detecting if the seen value is the first value in
the array (in which case the value in question would be the key and not
the value).