Support parsing files with missing brackets
There are many instances of game files that are missing a closing bracket:

```
HOI4\history\countries\GYA - Guyana.txt
HOI4\history\countries\NOR - Norway.txt
HOI4\history\units\FRA_1936.txt
...
```

Where the last object won't contain a closing bracket.

There's a dozen or so instances of this. Since it is so common, we
should fixup the mistake as act like a closing bracket was present.