Add `take_last` deserialization attribute
[Paperman][0] melted saves (ironman to normal) include every checksum
that is part of the original ironman save, even though the combined
format should only have a single checksum. In order to support parsing
paperman melted saves, which is how skanderbeg converts them, this
commit introduces a new attribute that will only take the last instance
of a field instead of throwing a duplicate field error.

```rust
pub struct Model {
    #[jomini(take_last)]
    checksum: String,
}
```

[0]: https://gitgud.io/nixx/paperman