Allow dates to be parsed from numbers
[Paperman][0] melted saves (ironman to normal) mistakenly write some
dates as numbers (like `1.1.1` as `43808760`). Since skanderbeg users
load up their paperman saves in rakaly, we should be able to parse these
dates masquerading as numbers.

Notably `RawDate` does not support parsing numbers as dates as that would
throw off the downstream users that use it as a heuristic to determine
if the given text is a date or number (in which case all non-paperman
saves data should be detected as numbers and not mistakenly written
dates)

[0]: https://gitgud.io/nixx/paperman