Remove WriteVisitor in favor of precision functions
EU4 ai section / file has a different level of precision for f32 data
than the rest of the save (6 decimal places instead of 3), so it doesn't
make sense for users to define a static `WriteVisitor` for the duration
of a `TextWriter` when the precision can change in a file.

In it's place are two new functions `write_f32_precision` and
`write_f64_precision` that take a second argument with the number of
decimal places to write out.