Allow keys to be written as quoted
Prior to this PR, the writer would check if the next entry should be a
key, and if it was: redirect the quote write to unquoted.

This was based on the assumption that keys always shouldn't be quoted.

It turns out that this assumption isn't true, as can be seen by EU4's
culture group entry:

```
  culture_group={
    "germanic"={
       # ...
    }
  }
```

For downstream users that would like to keep the old behavior, use the
`expecting_key` function to manually redirect to an unquoted write.