Unwrap result for json vec and string implementations
These don't need to return a result as we know the vec writer that is being
written to won't error, and the type being serialized won't error. Thus
it should be safe for us to unwrap the result so that the return type is
a bit more ergonomic for clients.