Fix stack overflow on flattened struct on input with headers
Using the `flatten` serde attribute is not recommended as it only works
with self describing formats. That said, if one happens to use the
attribute, it shouldn't cause a stack overflow when given an input with
headers (eg: `a = rgb { 10 20 30 }`).

Since headers are also not really supported with deserialization, this
is a bit of a hack of forcing forward movement in the reader.