Move examples to bin directory
It's annoying that the examples directory requires compilation of dev
dependencies. I was doing some benchmarking on a slower computer and I
much prefer to be able to run with `--no-default-features` and only need
to compile the actual crate + example. Saves about a minute each time a
need to do an iteration when benchmarking large data that doesn't fit in
the repo.

I also updated the output so that it showed elapsed time to better
exclude the time it takes to ingest stdin