Expose RawDate::from_binary for Vic3 dates
Time in Vic3 has an hour component that is either 0, 6, 12, or 18. It
feels a bit wrong to add such a specific time system to the repo (even
though stellaris is here). So I'm exposing `RawDate::from_binary` as it
is a missing primitive.

Perhaps later this date system can be added here. But I'd like to be
able to iterate on vic3 time in its own repo.