More binary parsing improvements based on data and fuzz equality
Realizing that there are 125k integers in arrays that weren't being
captured by the usual fast-path of key=value detection, I installed a
special loop just for i32 instances.

Also included a few more conditionals in the fast-path.

Together the throughput improvement is between 5-10% for EU4 saves. CK3
and VIC3 were unchanged.

This commit also adds in a mechanism for turning off the fast-path. This
is mainly used in fuzzing to ensure that both functions return the same
result.