Date parsing 30x perf improvement
Date parsing from text is a hot path of save file parsing. Even the
smallest EU4 saves will have a hundred thousand of dates. The largest saves
can crack a million.

When attempting to parse the history section of EU4, we attempt to parse
a lot of dates (yes, even for binary saves)

When profiling, I noticed that date parsing was
popping up so I decided to dive in and apply the following
optimizations:

- Add happy path detection that can do ascii number validation and
  parsing in two instructions.
- In the general path, we know that month, day, and hour can only be two
  digits long, so use specialized routines for parsing short numbers
- Add length checking for dates

This resulted in major improvements across the board for all use cases