Add json type narrowing configuration
When converting to JSON, values are eagerly narrowed from strings to
numbers or booleans. However, there are times where one may encounter a
field like:

```
a="01"
```

Current type narrowing will coalesce it to `1`, which is may be too
lossy for some.

This commit introduces a few configuration options for type narrowing:

- All (default). Current behavior of type narrowing all values
- Unquoted. Only type narrow values that aren't quoted
- None. Never type narrow