Add ondemand binary deserializer
I was doing some unrelated testing and saw that a 50MB binary file
resulted in a 200MB tape (and this is before any deserialization). This
fact didn't sit well with me and required an investigation.

The reason why so much memory is required is the number of tokens (each
16 bits) that occur in the binary file and how each one is much smaller
than the tape's intermediate representation (192 bits).

Simdjson's intermediate representation is 64 bits that can stretch to
128 bits for larger types. This seems possible to emulate and could cut
the weight of the intermediate representation by at least half.

But this got me thinking, is the intermediate representation really
pulling its weight? The native binary format is easy enough to parse
standalone that I realized that those who are fine with one-shot
deserialization could benefit from cutting out the middleman.

What do benchmarks show? The most favorable benchmarks show around a ~3x
throughput increase when deserializing a fraction of the input. If most
of the input is deserialized then the throughput increase drops to
around 50%.

This is a breaking change as deserialization error type has been changed
to `jomini::Error` instead of `jomini::DeserializeError` as the new
ondemand deserializer can encounter an early EOF (an error that didn't
exist previously in `DeserializeError` as it couldn't occur when
deserialization was after parsing).