Attempt deserialization for unexpected tokens
When a deserializer asks for an object to be deserialized, but the
current value isn't an object, deserialization would eagerly fail.

This commit changes this behavior by attempting to deserialize
current value, according to its type, even if it doesn't the request as
according to serde:

> The `deserialize_*` methods are hints to the deserializer for how to
> interpret the next piece of input
> [...]
>
> Note that a Deserializer will not necessarily follow the type hint, so
> the call to deserialize_i32 does not necessarily mean the Deserializer
> will call I32Visitor::visit_i32

So by not eagerly erroring, this opens use cases where the deserializer
can handle values that are strings or objects under a single enum
without resorting to an `untagged` enum which doesn't work well.