Support new binary i64 data type from HOI4 1.13
This new data type can be seen in HOI4 1.13 saves as
```
manpower = { ratio = 81900 }
```

This datatype technically conflicts with some detected tokens in
Imperator and Vic3 (`playbackrate`), but I believe this token to be
unused and shouldn't pose too much of a problem.

I determined that it was signed 64bit by editing the binary to turn all
bits on and I saw the output as -1.

This is a breaking change as `BinaryToken` is receiving an addition.